config: add new 'docstring-code-format' knob
This commit does the plumbing to make a new formatting option,
'docstring-code-format', available in the configuration for end users.
It is disabled by default (opt-in). It is opt-in at least initially to
reflect a conservative posture. The intent is to make it opt-out at
some point in the future.

The name was chosen due primarily because it serves as a
way to namespace other related options. For example, a
'docstring-code-line-length' option.